+++ tests/integration/test_parameters_report.py
@@ -16,6 +16,7 @@
 import pytest
 from unittest import mock
 
+from examples.params.test_binary_symbol_in_parameters import BINARY_TEXT
 from tests import REPORT_PORTAL_SERVICE
 from tests.helpers import utils
 
@@ -25,7 +26,8 @@ from tests.helpers import utils
     ('examples/test_simple.py', None),
     ('examples/params/test_in_class_parameterized.py', {'param': 'param'}),
     ('examples/params/test_different_parameter_types.py',
-     {'integer': 1, 'floating_point': 1.5, 'boolean': True, 'none': None})
+     {'integer': 1, 'floating_point': 1.5, 'boolean': True, 'none': None}),
+    ('examples/params/test_binary_symbol_in_parameters.py', {'text': BINARY_TEXT.replace('\0', '\\0')}),
 ])
 def test_parameters(mock_client_init, test, expected_params):
     """Verify different tests have correct parameters.